+++ t/03-test-connection.t
@@ -2,11 +2,15 @@
 
 use Modern::Perl;
 
-use Test::More tests => 2;
+use Test::More tests => 3;
+use Test::MockModule;
 
+use CGI;
 use File::Temp qw(tempdir);
+use YAML::XS   qw(Dump);
 
 use Koha::Database;
+use Koha::Encryption;
 use t::lib::TestBuilder;
 
 use Koha::Plugin::Com::ByWaterSolutions::PatronsImporterAdvanced;
@@ -52,3 +56,70 @@ subtest '_test_job_transport reports failure for an unknown file_transport_id' =
 
     $schema->storage->txn_rollback;
 };
+
+subtest 'configure() test branch reports a job with no file_transport_id instead of omitting it' => sub {
+    plan tests => 5;
+
+    $schema->storage->txn_begin;
+
+    my %template_params;
+    my $fake_template = FakeTemplate->new( \%template_params );
+
+    my $mock_base = Test::MockModule->new('Koha::Plugins::Base');
+    $mock_base->mock( 'get_template', sub { return $fake_template } );
+    $mock_base->mock( 'output_html',  sub { return } );
+
+    my $dir       = tempdir( CLEANUP => 1 );
+    my $transport = $builder->build_object(
+        {
+            class => 'Koha::File::Transports',
+            value => {
+                transport => 'local', download_directory => $dir, upload_directory => $dir,
+                host => '', port => 22, user_name => undef, password => undef, key_file => undef, passive => 1,
+            },
+        }
+    );
+
+    my $plugin = Koha::Plugin::Com::ByWaterSolutions::PatronsImporterAdvanced->new( { enable_plugins => 1 } );
+    $plugin->store_data(
+        {
+            configuration => Koha::Encryption->new->encrypt_hex(
+                Dump(
+                    [
+                        { name => 'No transport job', parameters        => { matchpoint => 'cardnumber' } },
+                        { name => 'Good job',         file_transport_id => $transport->id },
+                    ]
+                )
+            ),
+        }
+    );
+    $plugin->{cgi} = CGI->new('test=1');
+
+    $plugin->configure();
+
+    is( $template_params{test_completed}, 1, 'test branch ran to completion' );
+    my $results = $template_params{results};
+    is( scalar @$results, 2, 'a job with no file_transport_id still appears in the results' );
+    is( $results->[0]->{ok}, 0, 'the transportless job is reported as not ok' );
+    like( $results->[0]->{error}, qr/no file_transport_id/, 'its error explains the missing transport' );
+    ok( $results->[1]->{ok}, 'the properly configured job still tests ok' );
+
+    $schema->storage->txn_rollback;
+};
+
+{
+
+    package FakeTemplate;
+
+    sub new {
+        my ( $class, $params ) = @_;
+        return bless { params => $params }, $class;
+    }
+
+    sub param {
+        my ( $self, %args ) = @_;
+        $self->{params}->{$_} = $args{$_} for keys %args;
+    }
+
+    sub output { return q{} }
+}